+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+## 0.2.0 - 2016-12-29
+### Added
+- compile function
+- basic benchmarks script
 ## 0.1.0 - 2016-12-28
 ### Added
 - supdate macro
 - tests
 
-[Unreleased]: https://github.com/vvvvalvalval/supdate/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/vvvvalvalval/supdate/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/vvvvalvalval/supdate/compare/v0.1.0...v0.2.0